+++ R/env.R
@@ -556,6 +556,69 @@ env_inherits <- function(env, ancestor) {
   is_empty_env(env)
 }
 
+#' Lock an environment
+#'
+#' @description
+#'
+#' \badgeexperimental
+#'
+#' Locked environments cannot be modified. An important example is
+#' namespace environments which are locked by R when loaded in a
+#' session. Once an environment is locked it normally cannot be
+#' unlocked.
+#'
+#' Note that only the environment as a container is locked, not the
+#' individual bindings. You can't remove or add a binding but you can
+#' still modify the values of existing bindings. See
+#' [env_binding_lock()] for locking individual bindings.
+#'
+#' @param env An environment.
+#' @return The old value of `env_is_locked()` invisibly.
+#'
+#' @seealso [env_binding_lock()]
+#' @export
+#' @examples
+#' # New environments are unlocked by default:
+#' env <- env(a = 1)
+#' env_is_locked(env)
+#'
+#' # Use env_lock() to lock them:
+#' env_lock(env)
+#' env_is_locked(env)
+#'
+#' # Now that `env` is locked, it is no longer possible to remove or
+#' # add bindings. If run, the following would fail:
+#' # env_unbind(env, "a")
+#' # env_bind(env, b = 2)
+#'
+#' # Note that even though the environment as a container is locked,
+#' # the individual bindings are still unlocked and can be modified:
+#' env$a <- 10
+env_lock <- function(env) {
+  old <- env_is_locked(env)
+  lockEnvironment(env)
+  invisible(old)
+}
+#' @rdname env_lock
+#' @export
+env_is_locked <- function(env) {
+  environmentIsLocked(env)
+}
+
+#' Unlock an environment
+#'
+#' This function should only be used in development tools or
+#' interactively.
+#'
+#' @inheritParams env_lock
+#' @return Whether the environment has been unlocked.
+#'
+#' @keywords internal
+#' @export
+env_unlock <- function(env) {
+  invisible(.Call(rlang_env_unlock, env))
+}
+
 
 #' Scoped environments
 #'