+++ src/export/init.c
@@ -95,6 +95,8 @@ extern sexp* rlang_as_data_mask(sexp*, sexp*);
 extern sexp* rlang_data_mask_clean(sexp*);
 extern sexp* rlang_eval_tidy(sexp*, sexp*, sexp*);
 extern sexp* rlang_as_data_pronoun(sexp*);
+extern sexp* rlang_env_binding_are_promise(sexp*, sexp*);
+extern sexp* rlang_env_unlock(sexp*);
 
 // Library initialisation defined below
 sexp* rlang_library_load();
@@ -211,6 +213,8 @@ static const r_callable r_callables[] = {
   {"rlang_data_mask_clean",     (r_fn_ptr_t) &rlang_data_mask_clean, 1},
   {"rlang_eval_tidy",           (r_fn_ptr_t) &rlang_eval_tidy, 3},
   {"rlang_as_data_pronoun",     (r_fn_ptr_t) &rlang_as_data_pronoun, 1},
+  {"rlang_env_binding_are_promise", (r_fn_ptr_t) &rlang_env_binding_are_promise, 2},
+  {"rlang_env_unlock",          (r_fn_ptr_t) &rlang_env_unlock, 1},
   {NULL, NULL, 0}
 };